feat: change to typescript and add auth/signup api
