+++ src/app/api/auth/route.js
@@ -1,28 +1,26 @@
-import {connectDB, addData} from "@/lib/database.js";
-import UserModel from "@/models/UserSchema.js";
-import {loginUser, logoutUser, registerUser, getSession} from "@/lib/backend/authentication/authentication.js";
+// import {connectDB, addData} from "@/lib/database.js";
+// import UserModel from "@/models/UserSchema.js";
+// import {loginUser, logoutUser, registerUser, getSession} from "@/lib/backend/authentication/authentication.js";
 
-export async function GET(req){
-    await connectDB();
-    const sessionData = await getSession();
-    return new Response(JSON.stringify(sessionData), { 
-        headers: { 'Content-Type': 'application/json' } 
-    });
-}
+// export async function GET(req){
+//     await connectDB();
+//     const sessionData = await getSession();
+//     return new Response(JSON.stringify(sessionData), { 
+//         headers: { 'Content-Type': 'application/json' } 
+//     });
+// }
 
-export async function POST(req){
-    await connectDB();
-    const body = await req.json();
-
-    const {email, password} = body;
+// export async function POST(req){
+//     await connectDB();
+//     const body = await req.json();
     
-    try {
-        const result = await loginUser(body);
+//     try {
+//         const result = await loginUser(body);
 
-        return new Response(result);
-    } catch (error) {
-        console.log(error);
-        return new Response("Login Error");
-    }
-}
+//         return new Response(result);
+//     } catch (error) {
+//         console.log(error);
+//         return new Response("Login Error");
+//     }
+// }
