--- src/middleware.js
@@ -1,5 +1,0 @@
-import { updateSession } from "@/lib/backend/authentication/authentication.js";
-
-export async function middleware(request) {
-  return await updateSession(request);
-}
