--- src/models/UserSchema.js
@@ -1,20 +1,0 @@
-import mongoose from 'mongoose';
-
-const UserSchema = new mongoose.Schema({
-  username: {
-    type: String,
-    required: true,
-  },
-  email: {
-    type: String,
-    required: true,
-  },
-  password: {
-    type: String,
-    required: true,
-  },
-});
-
-const UserModel = mongoose.models?.User || mongoose.model('User', UserSchema);
-
-export default UserModel;
